Extract corner and shadow helper code to separate files so that cewe2pdf doesn't bloat too much Tidy the corner test mcf, tidy text box sizes in shadow mcf Fix a couple of really minor flake8 / pylint issues
